+++ spec/app/models/acolyte.rb
@@ -7,7 +7,7 @@ class Acolyte
   embeds_many :versions, as: :memorable
   belongs_to :church
 
-  default_scope asc(:name)
+  default_scope ->{ asc(:name) }
   scope :active, ->{ where(status: "active") }
   scope :named, ->{ where(:name.exists => true) }