+++ spec/app/models/appointment.rb
@@ -3,5 +3,5 @@ class Appointment
   field :active, type: Mongoid::Boolean, default: true
   field :timed, type: Mongoid::Boolean, default: true
   embedded_in :person
-  default_scope where(active: true)
+  default_scope ->{ where(active: true) }
 end