+++ spec/app/models/dog.rb
@@ -3,5 +3,5 @@ class Dog
   field :name, type: String
   has_and_belongs_to_many :breeds
   has_and_belongs_to_many :fire_hydrants, primary_key: :location
-  default_scope asc(:name)
+  default_scope ->{ asc(:name) }
 end