+++ spec/app/models/event.rb
@@ -17,6 +17,6 @@ class Event
     end
   end
 
-  scope :best, where(:kind.in => [ "party", "concert" ])
+  scope :best, ->{ where(:kind.in => [ "party", "concert" ]) }
   scope :by_kind, ->(kind){ where(:kind.in => [kind]) }
 end