+++ spec/app/models/person.rb
@@ -125,8 +125,8 @@ class Person
   accepts_nested_attributes_for :quiz
   accepts_nested_attributes_for :services, allow_destroy: true
 
-  scope :minor, where(:age.lt => 18)
-  scope :without_ssn, without(:ssn)
+  scope :minor, ->{ where(:age.lt => 18) }
+  scope :without_ssn, ->{ without(:ssn) }
   scope :search, ->(query){ any_of({ title: query }) }
 
   def score_with_rescoring=(score)