+++ spec/app/models/sound.rb
@@ -1,5 +1,5 @@
 class Sound
   include Mongoid::Document
   field :active, type: Mongoid::Boolean
-  default_scope where(active: true)
+  default_scope ->{ where(active: true) }
 end