+++ spec/app/models/symptom.rb
@@ -2,5 +2,5 @@ class Symptom
   include Mongoid::Document
   field :name, type: String
   embedded_in :person
-  default_scope asc(:name)
+  default_scope ->{ asc(:name) }
 end