+++ spec/mongoid/errors/invalid_scope_spec.rb
@@ -16,13 +16,13 @@ describe Mongoid::Errors::InvalidScope do
 
     it "contains the summary in the message" do
       expect(error.message).to include(
-        "Scopes in Mongoid must be either criteria objects or procs that wrap"
+        "Scopes in Mongoid must be procs that wrap"
       )
     end
 
     it "contains the resolution in the message" do
       expect(error.message).to include(
-        "Change the scope to be a criteria or proc wrapped critera."
+        "Change the scope to be a proc wrapped critera."
       )
     end
   end