+++ spec/mongoid/validatable/uniqueness_spec.rb
@@ -470,11 +470,11 @@ describe Mongoid::Validatable::UniquenessValidator do
 
           before do
             Dictionary.validates_uniqueness_of :name
-            Dictionary.default_scope(Dictionary.where(year: 1990))
+            Dictionary.default_scope(->{ Dictionary.where(year: 1990) })
           end
 
           after do
-            Dictionary.send(:strip_default_scope, Dictionary.where(year: 1990))
+            Dictionary.default_scoping = nil
             Dictionary.reset_callbacks(:validate)
           end